Add Career Paths blog post
Originally on the GDS Technology Blog. Interestingly, the original document is
no longer available. I haven't downloaded the new one, at the risk of that link
expiring as well, because I don't think it's that useful for another
organisation; when I no longer have access I will no longer need it, and it's
not my responsibility to ensure the GDS links are kept up to date.