Rename Games to Etc as a prelude to refactoring
I haven't added new games recently, and it's now actually a really small part of what I do, so I'd prefer this page to showcase all the stuff I want to share that's not my main job.

`games.html` is not a 404 but is no longer accessible via navigation. In a future commit I will redirect it to `etc.html`.