Redirect `games.html` to `etc.html`
Ideally this would be a 301 but as this site is GitHub pages I don't have any control over the server so instead using a meta refresh tag. Not ideal, but better than leaving a dead link.

[This article](https://superdevresources.com/redirects-jekyll-github-pages/) was useful.

I've styled the page as per the rest of the site for the back-up case where the redirect doesn't work, which at the moment involves duplication.