Change widths of two column layout
I'm going to change the whole site to two column layout (the only page that actually uses all three columns is the `cv.html` and it doesn't look great) so this commit changes it to be more how I would like it to look.

The only page using the two column layout at the moment is `etc.html`.