Use new includes for redirect layout
This requires passing two additional variables, the canonical link and the meta refresh. They can't be passed in inline and aren't available in the front matter, so this uses a liquid capture block to assign them to variables. It does mean there are blank lines in all the other pages, but that is acceptable.