Replace list of posts with a for loop of excerpts
This changes some things; for example the excerpt is just the first paragraph whereas previously I wrote some excerpts differently. However, in general I write the first paragraph to be a trailer, and the ones that are not like this won't be on this page anyway as I intend to limit it.

It also means that the links are to the post and from there to whatever it is if it's elsewhere; previously a post about, e.g. a talk, would have the link to the video from here. I don't think this is too problematic.