Make embedded iframes responsive
The code you copy from YouTube or Slideshare to embed video/slides into your website has a fixed height and width which means when included on the page it shrinks everything else to allow the embedded item to be displayed full width.

This change adds some CSS to handle these embedded items and removes the fixed heights and widths.