Move columns into preferred order for small screen
When the two column layout is not used, when a page width is less than 700px, the columns appear in the order they appear in source. This commit changes the order so the columns appear in the order I would like.

The only one that is now not how I would like it to appear is blog posts, which I would like to start with the content, not the menu. However, that is a larger change as the blog post layout inherits from the blog layout, so I will do that in a separate commit.