Remove 2 from style name
`left-2-column` and `right-2-column` referred to the name of the style when there was also a 3 column layout. There isn't now, so the 2 is redundant and possibly confusing.

It's still useful to have the position of the column in the name, especially now they may appear in the source in a different order.