Move blog layout contents to blog post layout
The blog post layout is no longer going to inherit from the blog layout as the columns are going to appear in a different order and that's pretty much all that's going on in the blog layout. This commit moves the contents of blog layout to here.

A later commit will reformat but leaving this like this to be clear what has changed.