Make list of posts appear after a post
On the JFDI page, on a mobile, it's fine for the top of the page to be a list of posts, starting with the most recent. However, if you have a link to a particular post, you don't want to have to scroll through 30+ links before you get to it. This changes the order of columns in the source so that the first thing you see is the post, and it is followed by the list of other posts.

On a screen with a width >= 700px, the list of posts will still appear to the left of the blog post, as it does on the JFDI page.