Don't float Twitter link right in small screens
Before this change, the link to me on Twitter was floated right in the heading part of the page. However, that looked really odd on a mobile, both because of the right float and also because of its positioning before the navigation menu.

This commit moves floating it to the right to inside the `@media` block so this will only happen on screens wider than 700px. It also moves the tag to be after the navigation menu so it doesn't appear before it on mobiles, and removes the top padding as this is now provided by the navigation div.