Stop name appearing too close together
I've no idea why this works as I can't see that it should have any effect. However, before this change, the line height was not 130% even though it said it was in the computed CSS. After this change, it is, which means on smaller screens it is correctly spaced out rather than on top of the next line.