Managed to change these settings
Jonathan Matthews reported that this affected his ability to sign up (https://twitter.com/jpluscplusm/status/927197893919952896) which is really annoying as it hadn't happened when I tested it. I went to edit the form to say "you only need to fill in the email address" and a new column had appeared on the right hand side which allowed me to delete the fields I didn't want. I'm not sure why that didn't appear before, or why it allowed me to do my test sign ups without entering any of those details, but hopefully it is fixed now.