Add margin to code block
This means that when a blog post finishes with a code block (e.g. https://www.annashipman.co.uk/jfdi/awk.html), there is white space after it and before the horizontal line that divides posts, so it's visually clearer that one post had finished and the next has started.