Make first paragraph markdown rather than HTML
The index page of my site automatically displays the first section. However, it's looking for a markdown break, so if the whole post is HTML (the easiest way to grab it from the GDS blog) the it includes the whole blog post as an excerpt.

This change separates out the first paragraph so that only that appears on the index page. It's not the best introductory paragraph but the title of the post is clear enough I think.