Add note about measurement to business case post
An oversight from this post was how useful it is to indicate how you would measure success. The business case can be used as part of the monitoring and evaluation of the service. In any case, it's a useful exercise to think about how success will be measured.

Since I still refer to this as reference, I'm adding it to the post.