Remove CV
I have just learned that at a certain level of seniority, you don't want your CV to be publically available, you want people to reach out to talk to you if they want to see it. Obviously, it remains available in the commit history for those who want to see the last publically available version!