Only show the latest 4 posts in full on JFDI
This page is too long to load on a phone or in some browsers, as it is all my posts end to end. Sadly I will have to only show the lastest 4.

People have told me they spend time reading articles on my blog so I need to work out how to make the rest of the articles available in that same way.