Convert relative links to absolute in atom feed
The mailing list uses the atom feed to send out the updates, but in the atom feed, and therefore the emails, relative links don't work. This affects both links and images. If you read atom in a web browser it works, but in the email it doesn't. This change replaces relative links with absolute ones just for the atom feed.