Style images for emails
The emails are generated from atom.xml which doesn't have styles and in any case email clients don't tend to read style sheets, so I will need to remember in future not to use markdown for images (or presumably anything else embedded like slides or video) and make sure it has at least `max-width: 100%`; hopefully this will stop the image appearing huge in emails in future.

This change won't make any difference to the email that's already been sent but might remind me to do this next time.